Refactoring.
- Introduce NotASharedDatabaseException and Optional for databaseID
- Reduce logic calls whithin the saveAs method for shared databases.
- Rename all occurrences databaseID to sharedDatabaseID
- Remove semantical nop
- Introduce clear* methods in OpenDatabaseAction
- Rename package
- Call startBackupTask() from initializing method.
- Remove unused imports. Change null-check.
- Add a check before deleting the .swp file

+++ src/main/java/net/sf/jabref/logic/autosaveandbackup/BackupManager.java
@@ -32,6 +32,8 @@ import org.apache.commons.logging.LogFactory;
 /**
  * Backups the given bib database file from {@link BibDatabaseContext} on every {@link BibDatabaseContextChangedEvent}.
  * An intelligent {@link ExecutorService} with a {@link BlockingQueue} prevents a high load while making backups and rejects all redundant backup tasks.
+ *
+ * This class does not manage the .bak file which is created when opening a database.
  */
 public class BackupManager {
 
@@ -72,8 +74,13 @@ public class BackupManager {
         }
     };
 
+
     @Subscribe
     public synchronized void listen(@SuppressWarnings("unused") BibDatabaseContextChangedEvent event) {
+        startBackupTask();
+    }
+
+    private void startBackupTask() {
         try {
             executor.submit(backupTask);
         } catch (RejectedExecutionException e) {
@@ -92,7 +99,9 @@ public class BackupManager {
         bibDatabaseContext.getMetaData().unregisterListener(this);
         executor.shutdown();
         try {
-            Files.delete(backupPath);
+            if (Files.exists(backupPath) && !Files.isDirectory(backupPath)) {
+                Files.delete(backupPath);
+            }
         } catch (IOException e) {
             LOGGER.error("Error while deleting the backup file.", e);
         }
@@ -112,6 +121,7 @@ public class BackupManager {
         if (originalFile.isPresent()) {
             backupManager.originalPath = originalFile.get().toPath();
             backupManager.backupPath = Paths.get(backupManager.originalPath.toString() + BACKUP_FILENAME_ENDING);
+            backupManager.startBackupTask();
             bibDatabaseContext.getDatabase().registerListener(backupManager);
             bibDatabaseContext.getMetaData().registerListener(backupManager);
             runningInstances.add(backupManager);
